requirements.txt
aiomysql==0.2.0
APScheduler==3.10.4
pyotp==2.9.0
python-dotenv==1.1.1
python-telegram-bot==22.3
qrcode==8.2
redis==5.0.1
